requirements.txt
# Document Processing
docling>=2.15.0
docling-core>=2.19.0

# RAG Framework
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-vector-stores-qdrant>=0.4.0
llama-index-vector-stores-postgres>=0.3.0
llama-index-graph-stores-neo4j>=0.4.0
llama-index-embeddings-huggingface>=0.4.0
llama-index-llms-ollama>=0.4.0

# Vector Database Clients
qdrant-client>=1.12.0
pgvector>=0.3.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0

# Knowledge Graph
neo4j>=5.26.0

# Embeddings
sentence-transformers>=3.3.0
FlagEmbedding>=1.2.0

# API & Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Visualization
streamlit>=1.41.0
plotly>=5.24.0
pyvis>=0.3.2

# Utilities
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.0
rich>=13.9.0
tqdm>=4.67.0
